requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pdfplumber>=0.10.4
langchain-core>=0.2.0
langchain-anthropic>=0.1.1
langchain-openai>=0.1.0
requests>=2.31.0